Move(#67): findMemberIdAfterSaveMember 위치 변경
